Defer loading tagging callback until after tags have been grabbed from database

+++ static/www/control/plan_interview_controller.js
@@ -1,5 +1,5 @@
 function plan_interview_controller($scope, $http, $mdDialog, $routeParams,
-                                   $mdMedia, $location, taggingService,
+                                   $rootScope, $location, taggingService,
                                    flaggingService, authService) {
 
     var interviewID = $routeParams.id;
@@ -132,25 +132,36 @@ function plan_interview_controller($scope, $http, $mdDialog, $routeParams,
         }
     };
 
-    $('#tagbox').on('beforeItemAdd', function(event) {
+    var beforeTagAdd = function(event) {
         event.itemValue = taggingService.countTag(event.item);
         event.itemText = event.item + " (" + event.itemValue + ")";
         $scope.taglist.push(event.item);
-    });
+    };
 
-    $('#tagbox').on('beforeItemRemove', function(event) {
-        authService.getUserToken(function(idToken) {
-            $http.delete('/tag/' + event.item
-                + '/interview/' + interviewID + "?idToken=" + idToken).success(function(created) {
+    var beforeTagRemove = function(event) {
+        if(interviewID > 0) {
+            authService.getUserToken(function(idToken) {
+                $http.delete('/tag/' + event.item
+                    + '/interview/' + interviewID + "?idToken=" + idToken).success(function(created) {
+                });
             });
-        });
+        }
         if($scope.taglist.indexOf(event.item) > -1) {
             $scope.taglist.splice($scope.taglist.indexOf(event.item), 1);
         }
+    };
+
+    $rootScope.$on('tagNotification', function() {
+        $('#tagbox').on('beforeItemAdd', beforeTagAdd);
+        $('#tagbox').on('beforeItemRemove', beforeTagRemove);
     });
 
     taggingService.resetTags();
     loadScreen();
+
+    $('#tagbox').on('beforeItemAdd', function(event) {});
+
+    $('#tagbox').on('beforeItemRemove', function(event) {});
 }
 
 function DialogController($scope, $mdDialog) {
